+++ src/components/refactor/atoms/LottieLoader/style.ts
