+++ src/components/refactor/atoms/LottieLoader/types.ts
