--- src/components/refactor/organism/index.ts
@@ -1,2 +1,0 @@
-export { default as Loader } from './Loader';
-export { default as LottieLoader } from './LottieLoader';
